requirements.txt
segmentation_models_pytorch
torchvision
albumentations
pandas
pytest
matplotlib
tensorboard
kaggle
kornia
lightning